修改从订单确认页面跳转地址管理页面的bug

+++ front/src/pages/userInfo/Address.vue
@@ -174,7 +174,9 @@ export default {
     this.getDatas();
   },
   beforeRouteEnter (to, from, next) {
-    sessionStorage.setItem('pathName',from.name);
+    if(from.name == 'OrderConfirm'){
+      sessionStorage.setItem('pathName',from.name);
+    }
     next();
   }
 };
